+++ modules/elasticache/variables.tf
@@ -1,8 +1,11 @@
+# Required
+
+variable "instance_type_services" {}
 variable "platform_instance_id" {}
 variable "subnet_id_1" {}
 variable "subnet_id_2" {}
-variable "instance_type_services" {}
-variable "instance_type_facebox" {}
+
+# Data Sources
 
 data "aws_subnet" "subnet_1" {
   id = "${var.subnet_id_1}"